Added the latest source files from the old branch, I2C and Timer Tests

+++ Flightcontroller/Src/receiver.c
@@ -20,7 +20,7 @@ void HAL_TIM_IC_CaptureCallback(TIM_HandleTypeDef* htim)
 		if (measured_time > 3000)channel_select_counter = 0;
 		else
 		{
-			channel[channel_select_counter] = (int ) (measured_time);
+			channel[channel_select_counter] = measured_time;
 			channel_select_counter++;
 		}
 	}
